small cygwin warning fixes

+++ tests/testgasnet.c
@@ -1,6 +1,6 @@
 /*  $Archive:: /Ti/GASNet/tests/testgasnet.c                              $
- *     $Date: 2004/06/21 05:15:42 $
- * $Revision: 1.15 $
+ *     $Date: 2004/06/26 10:41:58 $
+ * $Revision: 1.16 $
  * Description: General GASNet correctness tests
  * Copyright 2002, Dan Bonachea <bonachea@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -227,7 +227,7 @@ int main(int argc, char **argv) {
     for (i=0; i < 10; i++) {
       ALLAM_REQ(partner);
 
-      GASNET_BLOCKUNTIL(NUMREP() == (int)NUMHANDLERS_PER_TYPE*3*(i+1));
+      GASNET_BLOCKUNTIL((int)NUMREP() == (int)(NUMHANDLERS_PER_TYPE*3*(i+1)));
     }
 
     MSG("*** passed AM test!!");
